requirements.txt
pyrogram
TgCrypto
Pyromod
psutil
aiohttp
aiofiles
python-dotenv
sqlalchemy~=1.3.23
psycopg2-binary
feedparser
